requirements.txt
expiringdict==1.2.1
Flask==1.1.1
flask_restplus==0.13.0
Flask_SQLAlchemy==2.4.1
Flask-Docs==0.1.2
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pandas==1.3.5
PyYAML==6.0
SQLAlchemy==1.3.10
requests==2.22.0
shellescape==3.4.1
six==1.12.0
SQLAlchemy==1.3.10
tzlocal==2.0.0
urllib3==1.22
Werkzeug==0.16.0
pycrypto==2.6.1
PyMySQL==1.0.2